requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
boto3==1.40.17
botocore==1.40.17
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
comm==0.2.3
debugpy==1.8.16
decorator==5.2.1
diffusers==0.35.1
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.116.1
filelock==3.19.1
frozenlist==1.7.0
fsspec==2025.7.0
grpclib==0.4.8
h2==4.2.0
hf-xet==1.1.8
hpack==4.1.0
huggingface-hub==0.34.4
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.0
ipykernel==6.30.1
ipython==8.37.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
jupyter_client==8.6.3
jupyter_core==5.8.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
modal==1.1.3
mpmath==1.3.0
multidict==6.6.4
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
packaging==25.0
parso==0.8.5
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.4.0
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.32.0
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
pyzmq==27.0.2
regex==2025.7.34
requests==2.32.5
rich==14.1.0
s3transfer==0.13.1
safetensors==0.6.2
shellingham==1.5.4
sigtools==4.0.1
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.47.2
sympy==1.14.0
synchronicity==0.10.2
tokenizers==0.21.4
toml==0.10.2
torch==2.8.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.55.4
triton==3.4.0
typer==0.16.1
types-certifi==2021.10.8.3
types-toml==0.10.8.20240310
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
watchfiles==1.1.0
wcwidth==0.2.13
yarl==1.20.1
zipp==3.23.0
